+++ resources/views/addemployee/show.blade.php
@@ -75,7 +75,19 @@
                         &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
                         <td><a href="{{route('addemployee.edit',$employee->id)}}" class="btn btn-warning">Update</a></td>
                         
-                        <td>{!! Form::open(['method' => 'DELETE', 'route'=>['addemployee.destroy', $employee->id]]) !!}
+                        <script>
+
+                        function ConfirmDelete()
+                         {
+                           var x = confirm("Are you sure you want to delete?");
+                           if (x)
+                               return true;
+                             else
+                           return false;
+                        }
+                        </script>
+
+                        <td>{!! Form::open(['method' => 'DELETE', 'route'=>['addemployee.destroy', $employee->id],'onsubmit' => 'return ConfirmDelete()']) !!}
                                     {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
                                     {!! Form::close() !!}</td>
                            </tr>
